+++ extract_msg/appointment.py
@@ -2,6 +2,7 @@ from extract_msg import constants
 from extract_msg.attachment import Attachment
 from extract_msg.message_base import MessageBase
 
+
 class Appointment(MessageBase):
     """
     Parser for Microsoft Outlook Appointment files.