+++ extract_msg/dev_classes/message.py
@@ -46,7 +46,7 @@ class Message(olefile.OleFileIO):
                 prefix += '/'
         self.__prefix = prefix
         self.__prefixList = prefixl
-        
+
         if tmp_condition:
             filename = self._getStringStream(prefixl[:-1] + ['__substg1.0_3001'], prefix=False)
         if filename is not None:
@@ -97,20 +97,20 @@ class Message(olefile.OleFileIO):
             tmp = self._getStream(filename + '001E', prefix = False)
             return None if tmp is None else tmp.decode(self.stringEncoding)
 
-    def Exists(self, filename):
+    def exists(self, filename):
         """
         Checks if :param filename: exists in the msg file.
         """
         filename = self.fix_path(filename)
         return self.exists(filename)
-    
+
     def sExists(self, filename):
         """
         Checks if string stream :param filename: exists in the msg file.
         """
         filename = self.fix_path(filename)
         return self.exists(filename + '001F') or self.exists(filename + '001E')
-    
+
     def fix_path(self, filename, prefix=True):
         """
         Changes paths so that they have the proper
@@ -194,7 +194,7 @@ class Message(olefile.OleFileIO):
         except AttributeError:
             self._date = self._prop.date
             return self._date
-    
+
     @property
     def mainProperties(self):
         """
@@ -276,7 +276,7 @@ class Message(olefile.OleFileIO):
                 # Now, this next line SHOULD work, but it is possible that it might not...
                 self.__stringEncoding = str(enc)
                 return self.__stringEncoding
-    
+
     @stringEncoding.setter
     def stringEncoding(self, enc):
         self.__stringEncoding = enc